delete GitConfigurationRepository
This commit deletes the `GitConfigurationRepository` and related classes. It has been deprecated for a while. A planned new feature of configs with structured types would require adding more testing for this repository. Instead of making this effort, the repo is removed prior to developing this new feature instead.
