+++ src/test/java/org/zalando/baigan/proxy/ConfigurationServiceBeanFactoryIT.java
@@ -9,16 +9,12 @@ import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.FilterType;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.zalando.baigan.BaiganSpringContext;
-import org.zalando.baigan.annotation.BaiganConfig;
 import org.zalando.baigan.annotation.ConfigurationServiceScan;
-import org.zalando.baigan.context.SpringTestContext;
+import org.zalando.baigan.fixture.SomeConfiguration;
 import org.zalando.baigan.service.ConfigurationRepository;
-import org.zalando.baigan.service.github.GitConfig;
 
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -30,24 +26,19 @@ import static org.mockito.Mockito.when;
 @ContextConfiguration(classes = {ConfigurationServiceBeanFactoryIT.TestContext.class})
 public class ConfigurationServiceBeanFactoryIT {
 
-    @Configuration
-    @ComponentScan(
-            basePackageClasses = {BaiganSpringContext.class},
-            excludeFilters = @ComponentScan.Filter(
-                    classes = SpringTestContext.class,
-                    type = FilterType.ASSIGNABLE_TYPE))
+    @ComponentScan(basePackageClasses = {BaiganSpringContext.class})
     @ConfigurationServiceScan(basePackages = "org.zalando.baigan.proxy")
     static class TestContext {
 
         @Bean
-        ConfigurationRepository configurationRepository(final GitConfig configuration) {
+        ConfigurationRepository configurationRepository(final SomeConfiguration someConfiguration) {
             return mock(ConfigurationRepository.class);
         }
 
-        @Bean(name = "gitConfiguration")
-        GitConfig gitConfiguration() {
-            final GitConfig config = mock(GitConfig.class);
-            when(config.getGitHost()).thenReturn("raw.com");
+        @Bean(name = "someConfiguration")
+        SomeConfiguration someConfiguration() {
+            final SomeConfiguration config = mock(SomeConfiguration.class);
+            when(config.someValue()).thenReturn("a value");
             return config;
         }
 
@@ -72,8 +63,8 @@ public class ConfigurationServiceBeanFactoryIT {
 
         @Override
         public Object postProcessAfterInitialization(final Object bean, final String beanName) throws BeansException {
-            if ("gitConfiguration".equals(beanName)) {
-                when(((GitConfig) bean).getGitHost()).thenReturn("post-processed.com");
+            if ("someConfiguration".equals(beanName)) {
+                when(((SomeConfiguration) bean).someValue()).thenReturn("a post-processed value");
             }
             return bean;
         }
@@ -95,25 +86,19 @@ public class ConfigurationServiceBeanFactoryIT {
         }
     }
 
-    @BaiganConfig
-    public interface TestFeature {
-
-        Boolean enabled();
-    }
-
     @Autowired
-    private GitConfig gitConfig;
+    private SomeConfiguration someConfig;
 
     @Autowired
     private MyDependency myDependency;
 
     @Test
-    public void allowsPostProcessingOfBeans() throws Exception {
-        assertThat(gitConfig.getGitHost(), is("post-processed.com"));
+    public void allowsPostProcessingOfBeans() {
+        assertThat(someConfig.someValue(), is("a post-processed value"));
     }
 
     @Test
-    public void allowsPostProcessingOfFactoryBeans() throws Exception {
+    public void allowsPostProcessingOfFactoryBeans() {
         assertThat(myDependency, is(notNullValue()));
     }
 }
